requirements.txt
python-dotenv==1.0.1
synapseclient==4.6.1
nibabel==5.3.2
nilearn==0.11.1
matplotlib==3.10.0
numpy==1.26.4
torch==2.5.1+cu121
monai==1.4.0
torch-geometric==2.6.1
scikit-image==0.25.0
plotly==5.24.1
nbformat==5.10.4
scipy==1.14.1
